Updated fastqc workflow to follow the style guide.
The fastqc workflow now commits to running the complete workflow in the
execution folder used by cromwell. This ensures that the duplication strategy
can be easily implemented (more relevant when running on rivanna on a large
number of samples). In the end the output is copied to the final_out_dir.